% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolingR-package.R
\docType{package}
\name{biolingR-package}
\alias{biolingR}
\alias{biolingR-package}
\title{biolingR: Data On Biolinguistic Diversity}
\description{
It provides data on biological and linguistic diversity. As of now, it provides a data frame with number of languages and Biological Intactness Index by country.
}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID})

}
\keyword{internal}
